Create a separate tmux module for performace
The code called by tmux for the status bar no longer leads to the import
of many modules, which leads to a slow start-up time. So we have a
separate module with minimal imports which is much faster.

Use a context manager to handle the unshelving and shelving of
persistent data.

Simplify the tomato bash script, so that it directly passes command line
arguments to the main module.